+++ libraries/azure_network_security_groups.rb
@@ -1,33 +1,18 @@
 # frozen_string_literal: true
 
-require 'azurerm_resource'
+require 'azurerm_network_security_groups'
 
-class AzureNetworkSecurityGroups < AzurermResource
+class AzureNetworkSecurityGroups < AzurermNetworkSecurityGroups
   name 'azure_network_security_groups'
-  desc 'Verifies settings for Network Security Groups'
+  desc '[DEPRECATED] Please use azurerm_network_security_groups'
   example <<-EXAMPLE
     azure_network_security_groups(resource_group: 'example') do
       it{ should exist }
     end
   EXAMPLE
 
-  attr_reader :table
-
-  FilterTable.create
-             .add_accessor(:entries)
-             .add_accessor(:where)
-             .add(:exists?) { |obj| !obj.entries.empty? }
-             .add(:names, field: 'name')
-             .connect(self, :table)
-
   def initialize(resource_group: nil)
-    resp = client.network_security_groups(resource_group)
-    return if resp.nil? || (resp.is_a?(Hash) && resp.key?('error'))
-
-    @table = resp
-  end
-
-  def to_s
-    'Network Security Groups'
+    warn '[DEPRECATION] The `azure_network_security_groups` resource is deprecated and will be removed in version 2.0. Use the `azurerm_network_security_groups` resource instead.'
+    super
   end
 end
